refactor(message-core): export built avMessage
Remove message-angular to use subscribers from core
