--- packages/message-angular/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@availity/message-angular",
-  "private": true,
-  "version": "2.5.0",
-  "description": "angular wrapper for availity message",
-  "main": "src/index.js",
-  "author": "Kasey Powers <kasey.powers@availity.com>",
-  "license": "MIT",
-  "peerDependencies": {
-    "@availity/message-core": "^1.0.0-alpha.17",
-    "angular": "^1.6.4"
-  },
-  "devDependencies": {
-    "@availity/message-core": "^1.0.1",
-    "angular": "^1.6.4",
-    "angular-mocks": "^1.6.4"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
